admin/requirements.txt
django==3.2
flake8==4.0.1
python-dotenv==0.20.0
psycopg2-binary==2.9.3
django-split-settings==1.1.0
django-debug-toolbar==3.4.0
jinja2==3.1.2
celery==5.2.7
flower==1.0.0
SQLAlchemy==1.4.37
celery==5.2.7
django-celery-beat==2.3.0
django-createsuperuserwithpassword==2.0.0
grpcio-tools==1.47.0

hashids==1.3.1
sendgrid==6.9.7
websockets==10.3
orjson==3.6.7
pydantic==1.9.0